+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Changed
+- `JSONDeserializerTest`: modified to use `KType` values
+- `JSONSerializer`: optimisation of data classes
+- `JSONAuto`: added `stringify()`
 
 ## [0.3] - 2019-04-26
 ### Changed